requirements.txt
torch
torchvision
open_clip_torch
numpy
pillow
tqdm
scikit-learn
matplotlib
pandas
openai